Candidate elements are now set in the state vertex.
Candidate elements are set in the state vertex once the
CandidateElementExtractor has parsed the DOM. This way, you can see
which elements were fired once the crawl is down.

These elemets are CandidateElements not Eventables because if they are
not fired, they are not an Eventable (meaning a link between vertexes).
For that reason, the method `getUsedEventables()` was removed from
StateVertex. To be able to figure out which eventables were fired and
which weren't, you can retreived the edged going out of a state.

Fixes #350