Implemented the new DOM strippers
The idea is that you define zero or more of these strippers. They do not
have to produce a valid DOM, but they're there so you can influence how
DOMs are compared. StateVertexes are compared on the equals function of
these stripped doms so the more alike you make them, the less states
you'll have at the end of your crawl.