Added source and target states in CrawlOverview
Source and target states can now be shown in the CrawlOverview.
The state summary table displays a link next to the fan-in/-out
number to show all source/target states (if any).

The states are clickable, so this makes it especially easier to
diff against incoming states.

I modified some of the core classes: one typo and added a method
to get incoming states (in addition to the already existing outgoing
states).

Note: the states are shown ordered by name.